Remove old Sensors tab, rename sensor_config → sensors, rename file
- Delete old SensorsTab.vue (raw sensor data, moved to separate PR)
- Rename SensorConfigTab.vue → SensorsTab.vue
- Rename sensor_config key → sensors in registry, sidebar, tab-name,
  interval key, and CSS class
- Simplify SensorGraph.vue dual CSS selectors to single .tab-sensors
- Remove orphaned tabRawSensorData i18n key
