+++ src/components/sidebar/sidebar_items.js
@@ -7,7 +7,7 @@ export const sidebarItems = [
     { key: "flight_plan", mode: "disconnected", i18n: "tabFlightPlan", icon: "i-lucide-route", expert: true },
 
     { key: "setup", mode: "connected", i18n: "tabSetup", icon: "i-lucide-sliders-horizontal" },
-    { key: "sensor_config", mode: "connected", i18n: "tabSensorConfig", icon: "i-lucide-cpu" },
+    { key: "sensors", mode: "connected", i18n: "tabSensorConfig", icon: "i-lucide-cpu" },
     { key: "ports", mode: "connected", i18n: "tabPorts", icon: "i-lucide-cable" },
     { key: "configuration", mode: "connected", i18n: "tabConfiguration", icon: "i-lucide-settings" },
     { key: "power", mode: "connected", i18n: "tabPower", icon: "i-lucide-battery" },
@@ -29,7 +29,6 @@ export const sidebarItems = [
     { key: "osd", mode: "connected", i18n: "tabOsd", icon: "i-lucide-monitor", feature: "OSD" },
     { key: "vtx", mode: "connected", i18n: "tabVtx", icon: "i-lucide-radio-tower" },
     { key: "led_strip", mode: "connected", i18n: "tabLedStrip", icon: "i-lucide-lightbulb", feature: "LED_STRIP" },
-    { key: "sensors", mode: "connected", i18n: "tabRawSensorData", icon: "i-lucide-activity", expert: true },
     {
         key: "flight_plan_connected",
         tab: "flight_plan",
